edge/edge-api/requirements.txt
# IntelliOptics 2.0 - Edge API Dependencies

# Web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP client
httpx>=0.27.0  # Updated for IntelliOptics SDK compatibility
requests==2.31.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Azure integration (optional)
azure-identity==1.15.0
azure-storage-blob==12.19.0
azure-servicebus==7.11.4

# Configuration
python-dotenv==1.0.0
pyyaml==6.0.1
jinja2==3.1.2

# Image processing (for RTSP)
numpy<2  # OpenCV requires NumPy 1.x
opencv-python-headless==4.8.1.78
pillow==10.1.0

# Utilities
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
tenacity>=8.2.3
cachetools==5.3.2
ksuid==1.3

# Scheduling
APScheduler==3.10.4

# Monitoring
prometheus-client==0.19.0

# IntelliOptics SDK (now public repository)
git+https://github.com/thamain1/IntelliOptics-SDK@main
